+++ autodoc/__init__.py
@@ -1,0 +1,1 @@
+from .autodoc import autodoc
